move parsing stuff to parsing.scala
