+++ coulomb-parser/src/main/scala/coulomb/parser/QuantityParser.scala
@@ -18,170 +18,7 @@ package coulomb.parser
 
 import scala.util.{ Try, Success, Failure }
 
-import scala.util.parsing.combinator.{Parsers, RegexParsers}
-import scala.util.parsing.input.{NoPosition, Position, Reader}
-
 import coulomb._
-
-sealed class QuantityParserException(msg: String) extends Exception(msg)
-case class QPLexingException(msg: String) extends QuantityParserException(msg)
-case class QPParsingException(msg: String) extends QuantityParserException(msg)
-
-object lexer {
-  sealed trait UnitDSLToken
-  case class UNIT(unit: String) extends UnitDSLToken
-  case class PFUNIT(prefix: String, unit: String) extends UnitDSLToken
-  case class NUMLIT(n: Double, isInt: Boolean) extends UnitDSLToken
-  case object MULOP extends UnitDSLToken
-  case object DIVOP extends UnitDSLToken
-  case object POWOP extends UnitDSLToken
-  case object LPAREN extends UnitDSLToken
-  case object RPAREN extends UnitDSLToken
-
-  class UnitDSLLexer(units: Seq[String], pfunits: Seq[String]) extends RegexParsers {
-    override def skipWhitespace = true
-    override val whiteSpace = "[ \t]+".r
-
-    val unitRE = {
-      val t = (pfunits ++ units).mkString("|")
-      (s"($t)").r
-    }
-
-    val pfunitRE = {
-      val t1 = pfunits.mkString("|")
-      val t2 = units.mkString("|")
-      (s"($t1)($t2)").r
-    }
-
-    // order of "x.y" alternatives matters here, to bind longest possible sequence preferentially
-    val fpRE = "([+-]?(\\d+\\.\\d+|\\d+\\.|\\d+|\\.\\d+)([eE][+-]?\\d+)?)".r
-
-    def apply(expr: String): Try[List[UnitDSLToken]] = {
-      parse(tokens, expr) match {
-        case NoSuccess(msg, next) => scala.util.Failure(QPLexingException(msg))
-        case Success(result, next) => scala.util.Success(result)
-      }
-    }
-
-    lazy val tokens: Parser[List[UnitDSLToken]] = {
-      // pfunit before unit matters, to bind both prefix and unit preferntially
-      phrase(rep1(
-        numlit |
-        pfunit | unit |
-        mulop | divop | powop |
-        lparen | rparen
-      )) ^^ { x => x }
-    }
-
-    def unit: Parser[UNIT] = unitRE ^^ { u => UNIT(u) }
-
-    def pfunit: Parser[PFUNIT] = pfunitRE ^^ { case pfunitRE(pf, u) => PFUNIT(pf, u) }
-
-    def numlit: Parser[NUMLIT] = fpRE ^^ { t =>
-      Try { t.toInt } match {
-        case scala.util.Success(i) => NUMLIT(i, true)
-        case _ => NUMLIT(t.toDouble, false)
-      }
-    }
-
-    def mulop = "*" ^^ (_ => MULOP)
-    def divop = "/" ^^ (_ => DIVOP)
-    def powop = "^" ^^ (_ => POWOP)
-
-    def lparen = "(" ^^ (_ => LPAREN)
-    def rparen = ")" ^^ (_ => RPAREN)
-  }
-}
-
-object ast {
-  sealed trait UnitAST {
-    override def toString = this match {
-      case Quant(v, u) => s"(Rational($v)).withUnit[$u]"
-      case Unit(u) => u
-      case Mul(lhs, rhs) => s"%*[$lhs, $rhs]"
-      case Div(num, den) => s"%/[$num, $den]"
-      case Pow(rad, exp) => s"%^[$rad, $exp]"
-    }
-  }
-  case class Quant(v: Double, u: String) extends UnitAST
-  case class Unit(unitType: String) extends UnitAST
-  case class Mul(lhs: UnitAST, rhs: UnitAST) extends UnitAST
-  case class Div(num: UnitAST, den: UnitAST) extends UnitAST
-  case class Pow(rad: UnitAST, exp: Int) extends UnitAST
-
-  object UnitAST {
-  }
-}
-
-object parser {
-  import lexer._
-  import ast._
-
-  class UnitDSLParser(unitToType: Map[String, String]) extends Parsers {
-    override type Elem = UnitDSLToken
-
-    class UnitDSLTokenReader(tokens: Seq[UnitDSLToken]) extends Reader[UnitDSLToken] {
-      override def first: UnitDSLToken = tokens.head
-      override def atEnd: Boolean = tokens.isEmpty
-      override def pos: Position = NoPosition
-      override def rest: Reader[UnitDSLToken] = new UnitDSLTokenReader(tokens.tail)
-    }
-
-    def program: Parser[UnitAST] = {
-      phrase(quant)
-    }
-
-    def quant: Parser[UnitAST] = {
-      (qval ~ unitexpr) ^^ { case NUMLIT(v, _) ~ u => Quant(v, u.toString) }
-    }
-
-    def unitexpr: Parser[UnitAST] = {
-      exprL2
-    }
-
-    def exprL2: Parser[UnitAST] = {
-      val mul = (exprL1 ~ MULOP ~ unitexpr) ^^ { case lhs ~ _ ~ rhs => Mul(lhs, rhs) }
-      val div = (exprL1 ~ DIVOP ~ unitexpr) ^^ { case num ~ _ ~ den => Div(num, den) }
-      mul | div | exprL1
-    }
-
-    def exprL1: Parser[UnitAST] = {
-      val pow = (atomic ~ POWOP ~ exponent) ^^ { case b ~ _ ~ NUMLIT(e, true) => Pow(b, e.toInt) }
-      pow | atomic
-    }
-
-    def atomic: Parser[UnitAST] = {
-      val paren = (LPAREN ~ unitexpr ~ RPAREN) ^^ { case _ ~ expr ~ _ => expr }
-      val pfu = prefixunit ^^ { case PFUNIT(pname, uname) => Mul(Unit(unitToType(pname)), Unit(unitToType(uname))) }
-      val u = unit ^^ { case UNIT(uname) => Unit(unitToType(uname)) }
-      paren | pfu | u
-    }
-
-    def unit: Parser[UNIT] = {
-      accept("unit", { case u @ UNIT(uname) => u })
-    }
-
-    def prefixunit: Parser[PFUNIT] = {
-      accept("pfunit", { case pfu @ PFUNIT(pname, uname) => pfu })
-    }
-
-    def exponent: Parser[NUMLIT] = {
-      accept("exponent", { case exp @ NUMLIT(e, true) => exp })
-    }
-
-    def qval: Parser[NUMLIT] = {
-      accept("quantity value", { case qv @ NUMLIT(v, _) => qv })
-    }
-
-    def apply(tokens: Seq[UnitDSLToken]): Either[QPParsingException, UnitAST] = {
-      val reader = new UnitDSLTokenReader(tokens)
-      program(reader) match {
-        case NoSuccess(msg, next) => Left(QPParsingException(msg))
-        case Success(result, next) => Right(result)
-      }
-    }
-  }
-}
 
 class QuantityParser(qpp: coulomb.parser.infra.QPP[_]) {
   import scala.reflect.runtime.universe.{ Try => _, _ }
@@ -189,8 +26,8 @@ class QuantityParser(qpp: coulomb.parser.infra.QPP[_]) {
 
   private val toolbox = runtimeMirror(getClass.getClassLoader).mkToolBox()
 
-  val lex = new lexer.UnitDSLLexer(qpp.unames, qpp.pfnames)
-  val parse = new parser.UnitDSLParser(qpp.nameToType)
+  val lex = new coulomb.parser.lexer.UnitDSLLexer(qpp.unames, qpp.pfnames)
+  val parse = new coulomb.parser.parser.UnitDSLParser(qpp.nameToType)
 
   val unitDecls = qpp.decls.map { d => s"$d\n" }.mkString("")
